requirements.txt
absl-py==0.9.0
args==0.1.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
clint==0.5.1
colorama==0.4.3
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
floyd-cli==0.11.17
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
imageio==2.6.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Markdown==3.2
MarkupSafe==1.1.1
marshmallow==2.20.5
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pathlib2==2.3.5
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
tabulate==0.8.6
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0

